HomeworkSeven

+++ src/homeworkFive/Keyboard.java
@@ -2,7 +2,7 @@ package homeworkFive;
 
 public class Keyboard extends Product {
 
-    public Keyboard(int id, String name, double price){
+    public Keyboard(int id, String name, int price){
         super(id, name, price);
     }
 }
