+++ src/homeworkFive/Monitor.java
@@ -2,7 +2,7 @@ package homeworkFive;
 
 public class Monitor extends Product {
 
-    public Monitor(int id, String name, double price){
+    public Monitor(int id, String name, int price){
         super(id, name, price);
     }
 }
